Game variables adjusment and stats are now drawn

+++ GameVariables.cpp
@@ -9,7 +9,7 @@ RandomReal GameVariables::rand_sleep_interval(15, 60);
 RandomReal GameVariables::rand_spawn_interval(3, 10);
 
 long long GameVariables::PLASMA = 0;
-GLfloat GameVariables::PORTAL_HP = 100000;
+int GameVariables::PORTAL_HP = 100000;
 int GameVariables::MANA = 0;
 Window GameVariables::window{GameVariables::INIT_WINDOW_WIDTH,
                              GameVariables::INIT_WINDOW_HEIGHT};
